moved coupon and transferpair to dedicated files
